se maqueta una escructura básica
