+++ package.json
@@ -15,7 +15,8 @@
     "i18next": "^23.11.5",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
-    "react-i18next": "^14.1.2"
+    "react-i18next": "^14.1.2",
+    "react-router-dom": "^6.23.1"
   },
   "devDependencies": {
     "@tailwindcss/forms": "^0.5.7",
