+++ src/core/router/index.ts
@@ -1,0 +1,2 @@
+export * from "./router.component";
+export * from "./routes";
