+++ src/core/router/routes.ts
@@ -1,0 +1,7 @@
+interface SwitchRoutes {
+    root: string;
+}
+
+export const switchRoutes: SwitchRoutes = {
+    root: "/",
+}
