+++ src/pops/Header.tsx
@@ -4,6 +4,6 @@ import { useTranslation } from "react-i18next";
 export const Header: React.FC = () => {
     const { t } = useTranslation();
     return (
-        <div>{t('hola')}</div>
+        <header>{t('hola')}</header>
     )
 }
