+++ src/pops/SideBar.tsx
@@ -1,0 +1,10 @@
+import React from 'react'
+
+export const SideBar: React.FC = () => {
+
+    return (
+        <nav>
+            
+        </nav>
+    )
+}
