+++ src/scenes/home.scene.tsx
@@ -1,0 +1,12 @@
+import React from "react";
+import { LayoutApp } from "../layout/Layout";
+import { Header } from "../pops/Header";
+
+export const HomeScene: React.FC = () => {
+
+    return( 
+        <LayoutApp>
+            <Header />
+        </LayoutApp>
+    )
+}
