+++ src/scenes/index.tsx
@@ -1,0 +1,1 @@
+export * from "./home.scene"
